pyproject.toml
[project]
name = "embed-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "apache-airflow>=3.1.5",
    "bs4>=0.0.2",
    "linkedin-jobs-scraper>=5.0.2",
    "loguru>=0.7.3",
    "mcp>=1.21.0",
    "mysql-connector-python>=9.5.0",
    "openai>=2.8.0",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "uv>=0.9.10",
]
